chore: remove most of deprecated form components
