+++ src/components/form/components/styled-primitives/StyledFormElements.tsx
@@ -3,7 +3,7 @@ import { jsx, keyframes } from "@emotion/core"
 import React from "react"
 import { ThemeCss } from "../../../../theme"
 import { MdError } from "react-icons/md"
-import { FormGroupOptionsDirection } from "../FormGroupField"
+import { FormGroupOptionsDirection } from "../../types"
 
 const auxillaryTextCss: ThemeCss = theme => ({
   color: theme.colors.grey[50],
