+++ src/components/form/index.ts
@@ -1,27 +1,18 @@
-export * from "./components/FormField"
-export * from "./components/FormFieldBlock"
-
-export * from "./components/InputField"
 export * from "./components/InputFieldBlock"
 export * from "./components/InputConnectedField"
 
-export * from "./components/TextAreaField"
 export * from "./components/TextAreaFieldBlock"
 export * from "./components/TextAreaConnectedField"
 
-export * from "./components/SelectField"
 export * from "./components/SelectFieldBlock"
 export * from "./components/SelectConnectedField"
 
-export * from "./components/CheckboxField"
 export * from "./components/CheckboxFieldBlock"
 export * from "./components/CheckboxConnectedField"
 
-export * from "./components/CheckboxGroupField"
 export * from "./components/CheckboxGroupFieldBlock"
 export * from "./components/CheckboxGroupConnectedField"
 
-export * from "./components/RadioButtonField"
 export * from "./components/RadioButtonFieldBlock"
 export * from "./components/RadioButtonConnectedField"
 
@@ -35,3 +26,4 @@ export * from "./components/styled-primitives/StyledSelect"
 export * from "./components/styled-primitives/StyledTextArea"
 
 export * from "./styles"
+export * from "./types"
