+++ src/components/form/stories/CheckboxFieldBlock.stories.tsx
@@ -2,14 +2,13 @@
 import { jsx } from "@emotion/core"
 import React from "react"
 import { action } from "@storybook/addon-actions"
-import { CheckboxFieldBlock } from ".."
-import { FormFieldLabelSize } from "../components/FormField.helpers"
+import { CheckboxFieldBlock, StyledLabelSize } from ".."
 import { getFieldBlockSandboxProps } from "./stories.utils"
 import { text } from "@storybook/addon-knobs"
 import { withVariationsContainer } from "../../../utils/storybook"
 import CheckboxFieldBlockDocs from "./CheckboxFieldBlock.mdx"
 
-const LABEL_SIZES: FormFieldLabelSize[] = [`L`, `M`, `S`]
+const LABEL_SIZES: StyledLabelSize[] = [`L`, `M`, `S`]
 
 export default {
   title: `Form/Styled Blocks/CheckboxFieldBlock`,
