+++ src/components/form/stories/InputFieldBlock.stories.tsx
@@ -1,14 +1,13 @@
 /** @jsx jsx */
 import { jsx } from "@emotion/core"
 import { action } from "@storybook/addon-actions"
-import { InputFieldBlock, FormFieldBlockLayout } from ".."
-import { FormFieldLabelSize } from "../components/FormField.helpers"
+import { InputFieldBlock, FormFieldBlockLayout, StyledLabelSize } from ".."
 import { getFieldBlockSandboxProps } from "./stories.utils"
 import { text } from "@storybook/addon-knobs"
 import { withVariationsContainer } from "../../../utils/storybook"
 import InputFieldBlockDocs from "./InputFieldBlock.mdx"
 
-const LABEL_SIZES: FormFieldLabelSize[] = [`L`, `M`, `S`]
+const LABEL_SIZES: StyledLabelSize[] = [`L`, `M`, `S`]
 
 export default {
   title: `Form/Styled Blocks/InputFieldBlock`,
