+++ src/components/form/types.ts
@@ -1,0 +1,13 @@
+export type GroupControlProps = Omit<
+  React.ComponentPropsWithoutRef<"input">,
+  "value" | "name" | "id" | "aria-invalid" | "children"
+> & {
+  name: string // Force require "name" attribute
+  value: string // Force require "value" attribute
+}
+
+export type FormFieldLabelSize = "L" | "M" | "S"
+
+export type FormFieldBlockLayout = `horizontal` | `vertical`
+
+export type FormGroupOptionsDirection = `row` | `column`
